Update to newest version of Phaser, and add an example of interactive objects: transporters.
Transporters are defined in Tiled as objects (objects with name "transporter", and a custom property called "target" which is an object reference to the location that the transporter should jump to). Place a text label inside of the transporter area (with name "label" and whatever text you'd like). Then, when a user walks into that transporter area and presses the spacebar, they will be transported to the location that the transporter specifies as a target. See code in WorldMap.tsx added in this commit.